Refactor: Cleanup DataSource Protocol
I've removed some of the methods away from the protocol and just have them
as details of the core data implementation.

It became obvious when mocking the data source that it was exposing too much
and making life hard for tests. I also question why we need to hide the real
type away from the delegates, I don't think a lot of things are gained there
where as hiding the type from the view layer makes a lot of sense.

As a by product, also removed some forced casting, yay!