Refactor: Simplify Data Source
In hindsight, having the data source model the async actions was
wrong. The original thinking was to have the VC only rely on one
source of data for its rendering but it complicates the design
needlessly.

The network service should report pending results and the view
controller needs to be smart enough to use two plus deps to
work out what it needs to render.

This first part of the commit simplifies data sources.
Next need to get view controllers to render pending state.