Refactor: Centralising DetailDataSource Creation
Instead of having mutliple people constructing the datasource
for detail view segues, we create a registered service in the swinject
container. This factory service is then provided to the clients
wishing to setup detailviewcontrollers

As apart of this commit, we move the default implementation into
its own subclass. The protocol was too tied with coredata!
Atleast the subclass advertises the fact