Refactor: PaletteTableViewDelegate
- Aiming to make the delegates more testable
- Generally tried to use the type system to minimise undefined results

the willPresentDetail method is  a partial method that mutates detail view controller
but it only works in certain cases.

It doesn't work for non detail view controllers, so we specialise the type

It doesn't work when using a non palette datasource
this is harder to cleanly encode into the type system and still keep the tableview not
coupled to a specific model, just table rows. Chose to fatal error here as its
more of a developer mistake that will never work.

method now throws when unable to create the data source, this can happen for two reasons:
- index out of bounds
- problem syncing core data