Refactor: Better Seperation between Table data and view
Previously we had the table view directly bind to the datasource
the vc held. This was problematic for testing as it didn't really
allow us to easily mock out the datasource and test the vc.

This change is to have the table view binding indirectly via
the vc which now implements the required protocols.