Feature: User Onboarding
Infrastructure for onboarding user flow.

The main goal was to not leak details into the other vcs
and keep it encapsulated.

The only change we needed was to add the swinject assembler
into the default configuration for the object graph

The service is created and set to listen to root vc notification.
From there, it can grab the vc via the userinfo dict and get it to
present the onboarding VC.

We reuse the preference registery to keep a persistent flag and dictate
wether we should show the onboarding.