Feature: Swinject based feature flags for local testing
To better enable black box testing, we split the object graph
construction into parts which the user can switch on and off.

This allows the network and persistence layer to be easily mocked.
More so, it allows a fine grained feature inclusion mechanism.
The initial example is for disabling onboarding poping up on
startup.