Enhance: Making Palette Creation feature optional
The goal is to improve modularity in the application to the point
where any feature can be turned on and off via the app delegate
assembly sequence.

Here we make the main action VC conditionally include the action
IF the container can resolve the creation classes.

Likewise the delegate resolve isn't unwrapped so it can fail if required.